# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CDC Pipeline Control Plane API
  version: 1.0.0
  description: |
    HTTP control plane for the CDC (Change Data Capture) Pipeline platform.

    The control plane manages the lifecycle of CDC pipelines, source
    connections, and sink connections, all persisted in NATS KV. The
    worker process (`cmd/pipeline`) is a separate binary that consumes
    this configuration and runs the actual CDC pipelines; worker-internal
    types (messages, checkpoints, table stats, etc.) are intentionally
    NOT modeled in this spec.

    All routes are mounted under `/api/v1`.

    ## Authentication

    The `/login` endpoint is public and accepts a JSON body with
    `username` and `password` fields. On success it returns a JWT
    (HS256, 24h expiry) that must be sent on every subsequent request as:

        Authorization: Bearer <token>

    All other endpoints require a valid Bearer token. Missing or invalid
    tokens return `401`.
servers:
  - url: /api/v1
    description: Control plane API (same-origin)
...
security:
  - BearerAuth: []
paths:
  /login:
    post:
      tags: [Auth]
      summary: Authenticate and obtain a JWT
      description: |
        Exchanges username/password credentials for a signed JWT.
        This endpoint is the only public route; it does not require
        an `Authorization` header.
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /global:
    get:
      tags: [Global]
      summary: Get global default configuration
      description: Returns the cluster-wide default configuration for all pipelines.
      operationId: getGlobalConfig
      responses:
        '200':
          description: Global configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Global]
      summary: Update global default configuration
      description: |
        Replaces the global default configuration. The body is validated
        against the same rules as pipeline configuration timeouts.

        A `429` is returned if any pipeline is currently in a transitioning
        state (e.g. restarting); the server refuses to mutate global
        configuration while a transition is in flight because the change
        would be applied inconsistently across pipelines.
      operationId: updateGlobalConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalConfig'
      responses:
        '200':
          description: Updated global configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TransitionInProgress'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /stats/summary:
    get:
      tags: [Stats]
      summary: Get aggregated global stats summary
      description: |
        Returns pre-aggregated metrics across all pipelines. The server
        caches the summary for 30 seconds; on a stale or missing cache
        it recomputes on demand using a singleflight lock to avoid
        stampedes.
      operationId: getStatsSummary
      responses:
        '200':
          description: Aggregated stats summary
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: |
                  Free-form summary object. Fields include
                  `total_pipelines`, `healthy_count`, `error_count`,
                  `transitioning_count`, `total_rows_synced`,
                  `avg_lag_ms`. Modeled as `additionalProperties: true`
                  because the underlying `StatsSummary` type is a
                  worker-internal type not exposed in this spec.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /stats/history:
    get:
      tags: [Stats]
      summary: Get historical metrics
      description: |
        Returns time-series history points. Currently always returns an
        empty array; the time-series requirement has been removed.
      operationId: getStatsHistory
      responses:
        '200':
          description: History points (always empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                description: |
                  Always an empty array. Item shape mirrors the
                  worker-internal `HistoryPoint` type but is not modeled
                  here as a reusable schema.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /pipelines:
    get:
      tags: [Pipelines]
      summary: List all pipelines
      description: |
        Returns a paginated, optionally filtered list of pipelines.
        Query parameters:

        - `search` (optional): case-insensitive substring match against
          pipeline `id` and `name`.
        - `status` (optional): exact-match status filter. Computed from
          the pipeline's transition state and worker heartbeat.
        - `page` (optional, default `1`): 1-based page index.
        - `limit` (optional, default `10`): page size.
      operationId: listPipelines
      parameters:
        - in: query
          name: search
          required: false
          schema:
            type: string
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [healthy, error, transitioning, paused, stopped]
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: Paginated pipeline list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Pipelines]
      summary: Create a new pipeline
      description: |
        Creates a new pipeline configuration. The body is validated;
        every referenced `source` and `sink` ID must already exist
        in NATS KV or the request is rejected with `400`.

        A `429` is returned if the pipeline (by `id`) is currently in
        a transitioning state. The `started_at` field of the error
        response carries the timestamp at which the transition started.
      operationId: createPipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineConfig'
      responses:
        '201':
          description: Pipeline created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TransitionInProgress'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pipelines/{id}:
    parameters:
      - $ref: '#/components/parameters/PathID'
    get:
      tags: [Pipelines]
      summary: Get a single pipeline
      operationId: getPipeline
      responses:
        '200':
          description: >-
            Pipeline configuration merged with its computed status fields
            (see `Handler.pipelineWithStatus`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineListItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Pipelines]
      summary: Update an existing pipeline
      description: |
        Replaces the pipeline configuration. The path `id` overrides
        any `id` in the body. Referenced sources and sinks must exist.

        A `429` is returned if the pipeline is currently transitioning;
        the `started_at` field carries the transition start time.
      operationId: updatePipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineConfig'
      responses:
        '200':
          description: Updated pipeline configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TransitionInProgress'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Pipelines]
      summary: Delete a pipeline
      operationId: deletePipeline
      responses:
        '204':
          description: Pipeline deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pipelines/{id}/status:
    parameters:
      - $ref: '#/components/parameters/PathID'
    get:
      tags: [Pipelines]
      summary: Get live pipeline status and per-table stats
      description: |
        Returns an aggregated view of the pipeline's runtime state:

        - `status`: raw key/value snapshot of all NATS KV entries
          under the pipeline's status prefix (checkpoints, per-sink
          stats, transition state). Keys are NATS KV keys; values are
          their deserialized JSON. Modeled as
          `additionalProperties: true` because values can be any of
          several worker-internal types.
        - `tables`: aggregated per-table stats across all production
          (non-debug) sinks.
        - `sinks`: per-sink, per-table raw stats.
      operationId: getPipelineStatus
      responses:
        '200':
          description: Pipeline status payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pipelines/{id}/restart:
    parameters:
      - $ref: '#/components/parameters/PathID'
    post:
      tags: [Pipelines]
      summary: Trigger a pipeline restart
      description: |
        Writes a `PipelineTransitionState` entry to NATS KV to signal
        the worker to reload the pipeline. The endpoint returns `202`
        immediately; the actual restart is asynchronous.
      operationId: restartPipeline
      responses:
        '202':
          description: Restart accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pipelines/{id}/pause:
    parameters:
      - $ref: '#/components/parameters/PathID'
    post:
      tags: [Pipelines]
      summary: Pause a pipeline
      description: |
        Stops the pipeline from consuming while retaining its replication
        slot (WS-2, plan section 4.2 -- as opposed to `POST
        /pipelines/{id}/stop`, which drops the slot). Sets
        `desired_state` to `paused`, which triggers `ConfigManager`'s
        existing drain path (`stopWorker`/`honourDesiredState`); this
        endpoint does not drain the worker itself.

        Legal only when the pipeline's lifecycle state is `Running`; any
        other current state (already `Paused`, `Stopping`, ...) is
        rejected with `409` by the underlying transition table
        (`internal/protocol/lifecycle.go`).

        An optional `ttl` bounds how long the pause lasts before the
        auto-resume timer (WS-3) fires, capped at a 4h ceiling (OQ-3): a
        `ttl` above `4h0m0s` is rejected with `400`. Omitting `ttl` does
        *not* mean "pause indefinitely" -- it still gets a `paused_until`,
        defaulted to the same 4h ceiling, so an unbounded pause can never
        retain the replication slot (and accumulate WAL) forever. Either
        way it is stored as an absolute `paused_until` timestamp computed
        at request time, not a duration -- see plan section 8.

        The response may include a `warning` (plan section 5) projecting
        when this pause will exhaust the source's WAL budget, using the
        existing `cdc_source_slot_lag_bytes` probe's growth rate. Present
        only when that projection is shorter than the pause's effective
        TTL.
      operationId: pausePipeline
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PausePipelineRequest'
      responses:
        '200':
          description: Pipeline paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PausePipelineResponse'
        '400':
          description: >-
            Invalid request body, or `ttl` exceeds the maximum pause
            duration of `4h0m0s`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >-
            Illegal lifecycle transition (e.g. the pipeline is not
            currently Running).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pipelines/{id}/pause-projection:
    parameters:
      - $ref: '#/components/parameters/PathID'
    get:
      tags: [Pipelines]
      summary: Project a pause's time-to-breach before confirming it
      description: |
        Read-only: answers "if I paused for this `ttl` right now, would I
        hit the WAL budget guard first?" without taking any action -- no
        `desired_state` write, no lifecycle transition, callable from any
        current state. Plan section 5 is explicit that the projection must
        be shown "before the pause is confirmed"; `POST
        /pipelines/{id}/pause`'s own `warning` field is the same
        computation but only ever visible after the pause has already been
        committed. `PauseDialog` calls this endpoint as the operator adjusts
        the TTL slider/presets and re-renders the same warning text this
        would otherwise produce post-commit.

        Reuses the same `cdc_source_slot_lag_bytes` growth-rate sample and
        WAL-budget arithmetic as the pause endpoint (WS-3/WS-4); the two are
        computed from one shared helper so they cannot drift apart on what
        "the projection" means.
      operationId: pausePauseProjection
      parameters:
        - name: ttl
          in: query
          required: false
          schema:
            type: string
          description: >-
            Go duration string being considered (e.g. "2h"). Defaults to the
            4h ceiling when omitted, capped at `4h0m0s` -- the same bound
            `POST /pipelines/{id}/pause` enforces.
          example: 2h
      responses:
        '200':
          description: Projection for the given ttl
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PausePipelineProjectionResponse'
        '400':
          description: >-
            Invalid `ttl`, or `ttl` exceeds the maximum pause duration of
            `4h0m0s`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pipelines/{id}/start:
    parameters:
      - $ref: '#/components/parameters/PathID'
    post:
      tags: [Pipelines]
      summary: Start/resume a pipeline
      description: |
        Resumes a paused pipeline (WS-2, plan section 4.3: `Paused ->
        Resuming -> Running`), clearing `paused_until` and setting
        `desired_state` back to `running`, which triggers
        `ConfigManager`'s existing start path; this endpoint does not
        start the worker itself.

        The underlying transition table also accepts `start` from other
        states (`Stopped`, `NeedsResnapshot`, `Failed`), but those legs of
        the machine -- re-snapshot on resume, delete reconciliation --
        arrive in later workstreams. When called from one of those
        states, the lifecycle record still advances to whatever the
        transition table says, but `desired_state` is deliberately left
        unchanged so a pipeline that needs a re-snapshot is never handed
        to `ConfigManager` as a plain restart.
      operationId: startPipeline
      responses:
        '200':
          description: Pipeline lifecycle state after the transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineLifecycleRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Illegal lifecycle transition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pipelines/{id}/stop:
    parameters:
      - $ref: '#/components/parameters/PathID'
    post:
      tags: [Pipelines]
      summary: Stop a pipeline
      description: |
        Stops the pipeline from consuming and drops its replication slot,
        releasing WAL (WS-5, plan section 4.2 -- as opposed to `pause`,
        which retains the slot). Sets `desired_state` to `stopped`, which
        triggers `ConfigManager`'s existing drain path
        (`stopWorker`/`honourDesiredState`); the slot itself is dropped
        only after that drain completes
        (`internal/config/manager.go:finalizeStop`), so the response
        reflects the `Stopping` lifecycle state, not yet `Stopped` --
        poll `GET /pipelines/{id}` to observe the `Stopping -> Stopped`
        transition complete.

        Legal from `Running` or `Paused`; any other current state (already
        `Stopping`/`Stopped`, ...) is rejected with `409` by the underlying
        transition table (`internal/protocol/lifecycle.go`).

        Resuming a stopped pipeline (`POST /pipelines/{id}/start`) always
        requires a re-snapshot (`Stopped -> NeedsResnapshot -> Snapshotting`,
        WS-6) -- unlike resuming from `Paused`, which continues the
        existing snapshot in place.
      operationId: stopPipeline
      responses:
        '200':
          description: Pipeline lifecycle state after the transition (Stopping)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineLifecycleRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Illegal lifecycle transition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /sources:
    get:
      tags: [Sources]
      summary: List all source connections
      description: |
        Returns all source configurations. The `pass` field is
        decrypted before being returned.
      operationId: listSources
      responses:
        '200':
          description: Source list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Sources]
      summary: Create a new source connection
      description: |
        Creates a new source configuration. The `pass` field is
        encrypted at rest by the server before being persisted to
        NATS KV.
      operationId: createSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceConfig'
      responses:
        '201':
          description: Source created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sources/{id}:
    parameters:
      - $ref: '#/components/parameters/PathID'
    get:
      tags: [Sources]
      summary: Get a single source connection
      description: Returns the source configuration with `pass` decrypted.
      operationId: getSource
      responses:
        '200':
          description: Source configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Sources]
      summary: Update an existing source connection
      description: |
        Replaces the source configuration. The path `id` overrides
        any `id` in the body. The `pass` field is encrypted at rest.
      operationId: updateSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceConfig'
      responses:
        '200':
          description: Updated source configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Sources]
      summary: Delete a source connection
      operationId: deleteSource
      responses:
        '204':
          description: Source deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sources/{id}/schema:
    parameters:
      - $ref: '#/components/parameters/PathID'
    get:
      tags: [Sources]
      summary: List available schemas for a source
      description: |
        Queries pg_namespace on the source database live, excluding
        system schemas (pg_catalog, information_schema, pg_*).

        Note the whitelist semantic this feeds: an empty or omitted
        `schemas` list on a source means "public" ONLY, never "all
        schemas".
      operationId: getSourceSchema
      responses:
        '200':
          description: Available schemas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceSchemaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          description: Source database unreachable or schema discovery failed
  /sources/{id}/tables:
    parameters:
      - $ref: '#/components/parameters/PathID'
    get:
      tags: [Sources]
      summary: List discovered tables for a source
      description: |
        Returns all tables whose metadata has been discovered and
        persisted by the worker. Tables are filtered to those whose
        NATS KV key matches `*.sources.{id}.tables.*.metadata`.

        When no metadata has been persisted yet, the API falls back to
        querying `information_schema.tables` on the source directly. If
        that query fails — most commonly because the source is
        unreachable — the response is `502`, matching
        `GET /sources/{id}/schema`. It is deliberately not a `200` with
        an empty list, which made "cannot connect" indistinguishable
        from "connected, no tables".
      operationId: listSourceTables
      responses:
        '200':
          description: Discovered tables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceTablesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          description: Live table discovery against the source failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sources/test:
    post:
      tags: [Sources]
      summary: Test source connection
      description: |
        Tests connection to the source database using the provided configuration.
        If the password is blank and the ID matches an existing source, the
        existing password is used.
      operationId: testSourceConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceConfig'
      responses:
        '200':
          description: Connection test succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestConnectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: |
        HTTP Basic authentication. Declared for documentation/legacy
        purposes only; the actual `/login` endpoint accepts credentials
        in the JSON body, not via the HTTP Basic header.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Custom JWT issued by `/login`. Send as
        `Authorization: Bearer <token>` on every request to a protected
        route.
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication token (or bad credentials on /login)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    LoginRequest:
      type: object
      description: Credentials for `/login`. Both fields are required.
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: admin
    LoginResponse:
      type: object
      description: Successful login response.
      required: [token]
      properties:
        token:
          type: string
          description: HS256-signed JWT (24h expiry).
          example: eyJhbGciOiJIUzI1NiIs...
...